pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "askalot_qml"
version = "1.0.0"
description = "QML processing and analysis engine for Askalot platform"
authors = [
    {name = "Askalot Team", email = "team@askalot.io"}
]
license = {text = "Proprietary - All rights reserved"}
requires-python = "==3.12.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["qml", "questionnaire", "z3", "constraint-solving", "survey", "analysis"]
dependencies = [
    "Flask>=3.1.0",
    "PyYAML>=6.0.0",
    "jsonschema>=4.0.0",
    "z3-solver>=4.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-jsonschema",
]

[project.urls]
Homepage = "https://github.com/psaghelyi/askalot"
Repository = "https://github.com/psaghelyi/askalot"
Issues = "https://github.com/psaghelyi/askalot/issues"

[tool.hatch.build.targets.wheel]
packages = ["askalot_qml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests"
]

[tool.coverage.run]
source = ["askalot_qml"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-flask>=1.3.0",
]
